ctdb-doc: Document limitation of "ctdb reloadips"
This limitation can be removed with a reasonable amount of effort.  It
probably isn't worth doing until the public IP address configuration
is stored in a cluster-wide database.  That seems like the right time
to change the API to handle the details that
CTDB_CONTROL_GET_PUBLIC_IPS doesn't currently retrieve.

Signed-off-by: Martin Schwenke <martin@meltin.net>
Reviewed-by: Amitay Isaacs <amitay@gmail.com>